+++ ncrypt/pgpkey.c
@@ -187,7 +187,7 @@ static struct PgpKeyInfo *pgp_principal_key(struct PgpKeyInfo *key)
 static const char *pgp_entry_format_str(char *buf, size_t buflen, size_t col, int cols,
                                         char op, const char *src, const char *prec,
                                         const char *if_str, const char *else_str,
-                                        unsigned long data, MuttFormatFlags flags)
+                                        intptr_t data, MuttFormatFlags flags)
 {
   char fmt[128];
   bool optional = (flags & MUTT_FORMAT_OPTIONAL);
@@ -351,7 +351,7 @@ static void pgp_make_entry(char *buf, size_t buflen, struct Menu *menu, int line
 
   mutt_expando_format(buf, buflen, 0, menu->win_index->state.cols,
                       NONULL(C_PgpEntryFormat), pgp_entry_format_str,
-                      (unsigned long) &entry, MUTT_FORMAT_ARROWCURSOR);
+                      (intptr_t) &entry, MUTT_FORMAT_ARROWCURSOR);
 }
 
 /**
