+++ nntp/browse.c
@@ -51,7 +51,7 @@
 const char *group_index_format_str(char *buf, size_t buflen, size_t col, int cols,
                                    char op, const char *src, const char *prec,
                                    const char *if_str, const char *else_str,
-                                   unsigned long data, MuttFormatFlags flags)
+                                   intptr_t data, MuttFormatFlags flags)
 {
   char fn[128], fmt[128];
   struct Folder *folder = (struct Folder *) data;
