+++ remailer.c
@@ -446,7 +446,7 @@ static const char *mix_format_caps(struct Remailer *r)
 static const char *mix_format_str(char *buf, size_t buflen, size_t col, int cols,
                                   char op, const char *src, const char *prec,
                                   const char *if_str, const char *else_str,
-                                  unsigned long data, MuttFormatFlags flags)
+                                  intptr_t data, MuttFormatFlags flags)
 {
   char fmt[128];
   struct Remailer *remailer = (struct Remailer *) data;
@@ -519,7 +519,7 @@ static void mix_make_entry(char *buf, size_t buflen, struct Menu *menu, int num)
   struct Remailer **type2_list = menu->mdata;
   mutt_expando_format(buf, buflen, 0, menu->win_index->state.cols,
                       NONULL(C_MixEntryFormat), mix_format_str,
-                      (unsigned long) type2_list[num], MUTT_FORMAT_ARROWCURSOR);
+                      (intptr_t) type2_list[num], MUTT_FORMAT_ARROWCURSOR);
 }
 
 /**
